+++ RefEngine/Prims.h
@@ -1,19 +1,19 @@
 #pragma once
 
 #include "gl_core_4_4.h"
-
+#include "types.h"
 class Prims
 {
 public:
 	Prims() = delete;
 
-	static const GLfloat Triangle_Vertices[];
-	static const GLuint Triangle_NumberOfVerts;
+	static const float Triangle_Vertices[];
+	static const uint Triangle_NumberOfVerts;
 
-	static const GLfloat Cube_Vertices[];
-	static const GLushort Cube_Indices[];
-	static const GLfloat Cube_Colors[];
-	static const GLuint Cube_NumberOfVerts;
-	static const GLuint Cube_NumberOfIndices;
+	static const float Cube_Vertices[];
+	static const uint Cube_Indices[];
+	static const float Cube_Colors[];
+	static const uint Cube_NumberOfVerts;
+	static const uint Cube_NumberOfIndices;
 };
